git-svn-id: https://svn.apache.org/repos/asf/incubator/hama/trunk@703105 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hama/AbstractMatrix.java
@@ -153,10 +153,15 @@ public abstract class AbstractMatrix implements Matrix {
   }
 
   public void close() throws IOException {
-    admin.deleteTable(matrixName);
+    if(admin.isTableEnabled(matrixName)) {
+      admin.disableTable(matrixName);
+      admin.deleteTable(matrixName);
+    } else {
+      LOG.info("Table doesn't abled");
+    }
   }
 
-  public boolean save(String path) throws IOException {
-    return hAdmin.put(this.matrixName, path);
+  public boolean save(String name) throws IOException {
+    return hAdmin.put(this.matrixName, name);
   }
 }
